Fix min distance to objects not held by camera that is antiparallel to normals
