+++ assets/config/config.game.toml
@@ -23,10 +23,10 @@ translation_smoothing = 0.5
 rotation_smoothing = 0.5
 tracking_smoothing = 0.8
 min_degrees_looking_down = 10.0
-min_degrees_looking_up = 90.0
+min_degrees_looking_up = 30.0
 min_distance = 1.0
 max_distance = 10.0
 zoom_speed = 0.7
-min_distance_to_objects = 3e-1
+min_distance_to_objects = 4e-1
 zoom_in_smoothing = 0.2
 zoom_out_smoothing = 1.2
